storage: bucket_recv() should check rs lock
Locked replicaset (via config) should not allow any bucket moves
from or to the replicaset.

But the lock check was only done by bucket_send(). Bucket_recv()
allowed to receive a bucket even if the replicaset is locked. The
patch fixes it.

It didn't affect automatic bucket sends, because lock is
accounted by the rebalancer from the config. Only manual bucket
moves could have this bug.