Use fiber.clock() instead of .time() everywhere
Fiber.time() returns real time. It is affected by time corrections
in the system, and can be not monotonic.

The patch makes everything in vshard use fiber.clock() instead of
fiber.time(). Also fiber.clock function is saved as an upvalue for
all functions in all modules using it. This makes the code a bit
shorter and saves 1 indexing of 'fiber' table.

The main reason - in the future map-reduce feature the current
time will be used quite often. In some places it probably will be
the slowest action (given how slow FFI can be when not compiled by
JIT).

Needed for #147
Closes #246