error: introduce vshard.error.timeout()
The function returns a box.error.TIMEOUT error converted to the
format used by vshard.

Probably it wouldn't be needed if only Tarantool >= 1.10 was
supported - then error.make(box.error.new(box.error.TIMEOUT))
wouldn't be so bad. But 1.9 is supposed to work as well, and to
create a timeout error on <= 1.9 it is necessary to make a pcall()
which is long and ugly.

vshard.error.timeout() provides a version-agnostic way of
returning timeout errors.

The patch is motivated by timeout error being actively used in the
future patches about map-reduce.

Needed for #147